+++ Source/API/Refund.swift
@@ -59,14 +59,20 @@ public class Refund: NSObject {
     - Throws: LuhnValidationError judoID does not match the given length or is not luhn valid
     */
     init(receiptID: String, amount: Amount, paymentReference: String) throws {
+        // initialise variables
         self.receiptID = receiptID
         self.amount = amount
         self.paymentReference = paymentReference
         super.init()
         
+        // check if device is jailbroken and sdk was set to restrict access
+        if !Judo.allowJailbrokenDevices && Judo.isJailbroken() {
+            throw JudoError(.JailbrokenDeviceDisallowedError)
+        }
+        
         // luhn check the receipt id
         if !Session.isLuhnValid(receiptID) {
-            throw JudoError.LuhnValidationError
+            throw JudoError(.LuhnValidationError)
         }
     }
     
@@ -78,7 +84,7 @@ public class Refund: NSObject {
     
     - Returns: reactive self
     */
-    public func completion(block: (Response?, NSError?) -> ()) -> Self {
+    public func completion(block: (Response?, JudoError?) -> ()) -> Self {
         
         let parameters = Session.progressionParameters(self.receiptID, amount: self.amount, paymentReference: self.paymentReference)