+++ Source/Model/Reference.swift
@@ -36,6 +36,16 @@ public struct Reference {
     /// An object containing any additional data you wish to tag this payment with. The property name and value are both limited to 50 characters, and the whole object cannot be more than 1024 characters
     public let yourPaymentMetaData: [String : String]?
     
+    
+    /**
+     designated initialiser
+     
+     - parameter consumerRef: consumer reference string
+     - parameter paymentRef:  payment reference string
+     - parameter metaData:    meta data dictionary (defaults to nil)
+     
+     - returns: a Reference object
+     */
     public init(consumerRef: String, paymentRef: String, metaData: [String : String]? = nil) {
         self.yourConsumerReference = consumerRef
         self.yourPaymentReference = paymentRef